chore(rtdb): Refactor rtdb error logic
